requirements.txt
python-dotenv>=0.5.1
SQLAlchemy==1.4.34
psycopg2-binary
pandas==1.4.2
bs4
requests
html5lib
lxml
pytest
google-cloud-storage
google-cloud-bigquery
jinja2